requirements.txt
flask==3.0.3
flask-limiter==3.8.0
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==23.0.0
logtail-python==0.3.0
flask-cors==5.0.0
